More class per file goodness.  Less imports.
